Check compiler in test_package
