+++ recipes/dbg-macro/all/conanfile.py
@@ -16,19 +16,6 @@ class DbgMacroConan(ConanFile):
     def _source_subfolder(self):
         return "source_subfolder"
 
-    def configure(self):
-        minimal_cpp_standard = "11"
-        if self.settings.get_safe("compiler.cppstd"):
-            tools.check_min_cppstd(self, minimal_cpp_standard)
-
-        if self.settings.compiler == "gcc" and tools.Version(self.settings.compiler.version) < "5":
-            raise ConanInvalidConfiguration(
-                "dbg-mcro can't be used by {0} {1}".format(
-                    self.settings.compiler,
-                    self.settings.compiler.version
-                )
-            )
-
     def source(self):
         tools.get(**self.conan_data["sources"][self.version])
         extracted_dir = self.name + "-" + self.version
