+++ recipes/dbg-macro/all/test_package/conanfile.py
@@ -1,5 +1,6 @@
 import os
 from conans import ConanFile, CMake, tools
+from conans.errors import ConanInvalidConfiguration
 
 
 class TestPackageConan(ConanFile):
@@ -10,6 +11,19 @@ class TestPackageConan(ConanFile):
         cmake = CMake(self)
         cmake.configure()
         cmake.build()
+
+    def configure(self):
+        minimal_cpp_standard = "11"
+        if self.settings.get_safe("compiler.cppstd"):
+            tools.check_min_cppstd(self, minimal_cpp_standard)
+
+        if tools.Version(self.settings.compiler.version) < "5":
+            raise ConanInvalidConfiguration(
+                "dbg-mcro can't be used by {0} {1}".format(
+                    self.settings.compiler,
+                    self.settings.compiler.version
+                )
+            )
 
     def test(self):
         if not tools.cross_building(self.settings):
